Routes

+++ src/app/services/storage.service.ts
@@ -1,36 +1,44 @@
-import { Injectable } from '@angular/core';
-const USER_KEY ='auth-user';
+import { Injectable, PLATFORM_ID, Inject } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+
+const USER_KEY = 'auth-user';
 
 @Injectable({
   providedIn: 'root'
 })
 export class StorageService {
-  constructor() {}
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
 
   clean(): void {
-    window.sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.clear();
+    }
   }
 
   public saveUser(user: any): void {
-    window.sessionStorage.removeItem(USER_KEY);
-    window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.removeItem(USER_KEY);
+      window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    }
   }
 
   public getUser(): any {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return JSON.parse(user);
+    if (isPlatformBrowser(this.platformId)) {
+      const user = window.sessionStorage.getItem(USER_KEY);
+      if (user) {
+        return JSON.parse(user);
+      }
     }
-
     return {};
   }
 
   public isLoggedIn(): boolean {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return true;
+    console.log("Has entrado funcion");
+    if (isPlatformBrowser(this.platformId)) {
+      let user = window.sessionStorage.getItem(USER_KEY);
+      console.log(user);
+      return !!user;
     }
-
     return false;
   }
 }
